fix(cross-platform): macOS build compat + deletion API alignment for v0.1.0 release
- Fix statvfs type mismatch on macOS (u32 vs u64) via explicit u64::from casts
- Suppress unused_mut warning for collect_process_info on non-Linux platforms
- Suppress unused urgency variable on non-Linux notification paths
- Add FsStatsCollector import and fix emergency command pressure_check call site
- Include concurrent agent changes: deletion executor path-aware should_skip,
  guardrails refinements

+++ src/platform/pal.rs
@@ -132,11 +132,11 @@ impl Platform for LinuxPlatform {
             path: path.to_path_buf(),
             details: error.to_string(),
         })?;
-        let fragment = stat.fragment_size();
+        let fragment = u64::from(stat.fragment_size());
         Ok(FsStats {
-            total_bytes: stat.blocks().saturating_mul(fragment),
-            free_bytes: stat.blocks_free().saturating_mul(fragment),
-            available_bytes: stat.blocks_available().saturating_mul(fragment),
+            total_bytes: u64::from(stat.blocks()).saturating_mul(fragment),
+            free_bytes: u64::from(stat.blocks_free()).saturating_mul(fragment),
+            available_bytes: u64::from(stat.blocks_available()).saturating_mul(fragment),
             fs_type: mount.fs_type.clone(),
             mount_point: mount.path.clone(),
             is_readonly: stat.flags().contains(nix::sys::statvfs::FsFlags::ST_RDONLY),
